Change page title position a little
Because the first letter of the title is capital P and because of some
of weirdness in Lato, the heading is visibly not aligned with the rest
of the text on the page. Moving it somewhere between 2 and 3 pixels to
the left almost does the trick which is enough for now.