Add .htaccess to show expected server behavior
We expect the hosting server to comply with Chrome’s HSTS Preload List
Submission† requirements for HTTP Strict Transport Security (HSTS). We
expect that served disable certain potentially dangerous browser
features, namely:

* MIME type sniffing
* embedding our pages in other pages (in frames, iframes, etc.)
* reflected cross-site scripting (XSS) attacks

We also expect that HTML pages are requested via URLs without file
extensions. Lastly, we expect the server to redirect requests from
non-canonical hosts to a canonical one which is important for SEO.

We used WebPageTest‡ for guidance. Although this doesn’t completely
resolve all the issues identified by that particular test. We simply
focused on security in this config.

You can use the config as-is on Apache HTTP Server if the repository
content is placed in server’s or virtual host’s document root. The
directory context must allow override of the following types of directives:

* FileInfo
* Options – at least MultiViews.

It is also expected that mod_headers, mod_rewrite, and mod_negotiation
are loaded.

† hstspreload.org
‡ webpagetest.org