Stop relying on content negotiation in .htaccess
Sadly, Apache HTTP Server module mod_negotiation changes REQUEST_URI so
that it contains result of the negotiation. Therefore redirects to a
canonical host potentially have different path than the original
request. Specifically in our case, page URLs do not contain .html path
suffix, but a redirect after content negotiation for requested page does
contain such suffix. There seems to be no way around this, mainly
because mod_rewrite does not allow to turn off greedy matching of
regular expression which would allow us to safely remove the suffix for
redirect.

Without content negotiation, we try to do the poor-man’s equivalent with
mod_rewrite. We look at Accept HTTP header if it contains text/html, or
if it is set */* (presumably a default of many HTTP clients), or if it
is empty. In all these case and if the requested resource does not
exist, we check if there is one with .html file extension. If it is, we
serve it. This correctly sets the Vary HTTP header, so caching still
works.