Safely set HTTP header in .htaccess
This is inspired by Nextcloud’s approach.† We want to prevent duplicate
headers because Apache HTTP Server basically keeps two independent
tables with headers: onsuccess and always. We have to use “always” for
HSTS and no-sniff headers because without it, mod_rewrite ignores the
header. But there is a possibility that something else sets the same
header in “onsuccess” table.

We do not have to care about other headers because we do not use
“always” table for them. Firstly, it’s probably OK to set them only for
successful requests. Secondly, “always” table is also used to override
headers set by scripts which we do not have which is why we are not
using it.

† https://github.com/nextcloud/server/pull/19002