Prevent embedding of our pages
We added additional CSP directive – frame-ancestors. We set it to none
which has the same effect as setting HTTP header X-Frame-Options to DENY
which we already do. Unlike X-Frame-Options which is basically obsolete
and no longer recognized by many modern browsers, the CSP directive
frame-ancestors is standards-based and well supported. This was
suggested by webhint.io.